Change quantity to decimal type so it can support KG product as well

+++ VoyagerPos/Article/IArticle.cs
@@ -7,6 +7,6 @@
             get;
         }
 
-        decimal CalculatePrice(int quantity);
+        decimal CalculatePrice(decimal quantity);
     }
 }
